+++ modules/dggs/src/s2.ts
@@ -2,5 +2,14 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-export {S2Decoder} from './s2-decoder';
-export {getS2ChildIndex} from './s2-geometry/s2-token';
+export {S2Decoder, getS2Bounds, getS2LngLat, getS2BoundaryFlat} from './s2-decoder';
+export {
+  getS2ChildIndex,
+  getS2IndexFromToken,
+  getS2Level,
+  getS2TokenFromIndex,
+  isS2IndexValid,
+  isS2TokenValid
+} from './s2-geometry/s2-token';
+export {getS2DescendantIndex, getS2IndexFromCell} from './s2-geometry/s2-geometry';
+export type {S2Cell} from './s2-geometry/s2-geometry';